mhs2-bot/requirements.txt
python-dotenv==0.18.0
python-telegram-bot==12.7